pyproject.toml
[build-system]
requires = [
    "setuptools>=61.0",
    "wheel"
]
build-backend = "setuptools.build_meta"

[project]
name = "tokyo-sports-scraper"
version = "0.1.0"
description = "東京スポーツのオートレース情報をスクレイピングするライブラリ"
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
license-files = ["LICENSE"]
authors = [
    { name="shimomo", email="github@shimomo.net" }
]
dependencies = [
    "beautifulsoup4",
    "lxml",
    "python-dateutil",
    "requests"
]
